Added Stream.readExactStr, to preserve old semantics
The original implementation of MsgStream (before it was a subclass of
Stream) had `readStr` procs that raised an IOError if the entire
string couldn't be read. This does not match the behavior of the
standard Stream module, but the code here depends on it, and the
lack of this check broke a unit test.

(Note: This meant MsgPack had different behavior when using a regular
stream than when using a MsgStream. In the former case, some truncated
messages wouldn't be detected during unpacking, and bad data would be
returned.)

I created a Stream.readExactStr function that has the behavior the
code here expects, and changed all the calls to `readStr` to
`readExactStr`.